Get Plot Title

Get the title of the plot, or a default value.

get_plot_title(plot, valid_filename = TRUE, default = NA_character_)

Arguments

plot
a ggplot2 plot

valid_filename
a logical to indicate whether the returned value should be a valid filename, defaults to TRUE

default
the default value, if a plot title is absent

Examples

p <- plot2(mtcars, title = "Plotting **mpg** vs. **cyl**!")
#> i Using x = mpg
#> i Using y = cyl
#> Loading required namespace: AMR
#> i Using type = "point" since both axes are numeric
get_plot_title(p)
#> [1] "plotting_mpg_vs_cyl"
get_plot_title(p, valid_filename = FALSE)
#> [1] "Plotting **mpg** vs. **cyl**!"

p <- plot2(mtcars)
#> i Using x = mpg
#> i Using y = cyl
#> i Using type = "point" since both axes are numeric
# default is date and time:
get_plot_title(p)
#> [1] NA