Methods for plot2()

These are the implemented methods for different S3 classes to be used in plot2(). Since they have an extensive list of arguments, they are placed here on a separate manual page.

# S3 method for default
plot2(
  .data,
  x = NULL,
  y = NULL,
  category = NULL,
  facet = NULL,
  geom = NULL,
  x.title = NULL,
  y.title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "certe",
  colour_fill = NULL,
  x.lbl_angle = 0,
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x %n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x %n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x %n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0.5,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 10,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0.25,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.midpoint = NULL,
  category.trans = "identity",
  x.sort = NULL,
  category.sort = TRUE,
  facet.sort = TRUE,
  datalabels = TRUE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.format = "%n",
  datalabels.colour = "grey25",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ifelse(decimal.mark == ",", ".", ","),
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = FALSE,
  horizontal = FALSE,
  reverse = horizontal,
  smooth = FALSE,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = NULL,
  linetype = 1,
  bins = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "top",
  legend.title = NULL,
  legend.reverse = FALSE,
  legend.barheight = 6,
  legend.barwidth = 1.5,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = "Calibri",
  theme = theme_minimal2(),
  markdown = TRUE,
  x.category = NULL,
  y.category = NULL,
  ...
)

# S3 method for numeric
plot2(
  .data,
  x = NULL,
  y = NULL,
  category = NULL,
  facet = NULL,
  geom = "geom_point",
  x.title = NULL,
  y.title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "certe",
  colour_fill = NULL,
  x.lbl_angle = 0,
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x %n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x %n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x %n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0.5,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 10,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0.25,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.midpoint = NULL,
  category.trans = "identity",
  x.sort = NULL,
  category.sort = TRUE,
  facet.sort = TRUE,
  datalabels = FALSE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.format = "%n",
  datalabels.colour = "grey25",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ifelse(decimal.mark == ",", ".", ","),
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = FALSE,
  horizontal = FALSE,
  reverse = horizontal,
  smooth = FALSE,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = NULL,
  linetype = 1,
  bins = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "top",
  legend.title = NULL,
  legend.reverse = FALSE,
  legend.barheight = 6,
  legend.barwidth = 1.5,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = "Calibri",
  theme = theme_minimal2(),
  markdown = TRUE,
  x.category = NULL,
  y.category = NULL,
  ...
)

# S3 method for freq
plot2(
  .data,
  x = .data$item,
  y = .data$count,
  category = NULL,
  facet = NULL,
  geom = NULL,
  x.title = "Item",
  y.title = "Count",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "certe",
  colour_fill = NULL,
  x.lbl_angle = 0,
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x %n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x %n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x %n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0.5,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 10,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0.25,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.midpoint = NULL,
  category.trans = "identity",
  x.sort = "freq-desc",
  category.sort = TRUE,
  facet.sort = TRUE,
  datalabels = TRUE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.format = "%n",
  datalabels.colour = "grey25",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ifelse(decimal.mark == ",", ".", ","),
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = FALSE,
  horizontal = FALSE,
  reverse = horizontal,
  smooth = FALSE,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = NULL,
  linetype = 1,
  bins = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "top",
  legend.title = NULL,
  legend.reverse = FALSE,
  legend.barheight = 6,
  legend.barwidth = 1.5,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = "Calibri",
  theme = theme_minimal2(),
  markdown = TRUE,
  x.category = NULL,
  y.category = NULL,
  ...
)

# S3 method for sf
plot2(
  .data,
  x = NULL,
  y = NULL,
  category = NULL,
  facet = NULL,
  geom = NULL,
  x.title = NULL,
  y.title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "grey25",
  colour_fill = c("white", "certeblauw"),
  x.lbl_angle = 0,
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x %n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x %n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x %n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 10,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.midpoint = NULL,
  category.trans = "identity",
  x.sort = NULL,
  category.sort = TRUE,
  facet.sort = TRUE,
  datalabels = TRUE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.format = NULL,
  datalabels.colour = "black",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ifelse(decimal.mark == ",", ".", ","),
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = FALSE,
  horizontal = FALSE,
  reverse = horizontal,
  smooth = FALSE,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = 0.1,
  linetype = 1,
  bins = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "right",
  legend.title = TRUE,
  legend.reverse = FALSE,
  legend.barheight = 6,
  legend.barwidth = 1.5,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = "Calibri",
  theme = theme_minimal2(panel.grid.major = element_blank(), panel.grid.minor =
    element_blank(), panel.border = element_blank(), axis.title = element_blank(),
    axis.text = element_blank(), axis.line = element_blank(), axis.ticks =
    element_blank()),
  markdown = TRUE,
  x.category = NULL,
  y.category = NULL,
  ...
)

# S3 method for data.frame
plot2(
  .data,
  x = NULL,
  y = NULL,
  category = NULL,
  facet = NULL,
  geom = NULL,
  x.title = NULL,
  y.title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "certe",
  colour_fill = NULL,
  x.lbl_angle = 0,
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x %n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x %n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x %n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0.5,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 10,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0.25,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.midpoint = NULL,
  category.trans = "identity",
  x.sort = NULL,
  category.sort = TRUE,
  facet.sort = TRUE,
  datalabels = TRUE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.format = "%n",
  datalabels.colour = "grey25",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ifelse(decimal.mark == ",", ".", ","),
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = FALSE,
  horizontal = FALSE,
  reverse = horizontal,
  smooth = FALSE,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = NULL,
  linetype = 1,
  bins = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "top",
  legend.title = NULL,
  legend.reverse = FALSE,
  legend.barheight = 6,
  legend.barwidth = 1.5,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = "Calibri",
  theme = theme_minimal2(),
  markdown = TRUE,
  x.category = NULL,
  y.category = NULL,
  ...
)

Arguments

.data
data to plot

x
plotting 'direction': the x axis

y
values to use for plotting along the y axis

category
plotting 'direction': the category (called 'fill' and 'colour' in ggplot2)

facet
plotting 'direction': the facet

geom
type of visualisation to use, supports all ggplot2 geoms. It will be determined automatically if left blank.

In ggplot2, 'bars' and 'columns' are equal, while it is common to many people that bars are horizontal and columns are vertical. For this reason, geom = "bar" will set geom = "col" and horizontal = TRUE.

There is one special case: the shortcut geom = "barpercent", which will set geom = "col" and horizontal = TRUE and x.max_items = 10 and x.sort = "freq-desc and datalabels.format = "%n (%p)".

x.title
text to show on the x asis

y.title
text to show on the y asis

title
title to show

subtitle
subtitle to show

caption
caption to show

tag
tag to show

title.linelength
maximum number of characters per line in the title, before a linebreak occurs

title.colour
text colour of the title

subtitle.linelength
maximum number of characters per line in the subtitle, before a linebreak occurs

subtitle.colour
text colour of the subtitle

na.replace
character to put in place of NA values if na.rm = FALSE

na.rm
remove NA values from showing in the plot

facet.position
text

facet.fill
text

facet.bold
text

facet.italic
text

facet.size
text

facet.margin
text

facet.repeat_lbls_x
text

facet.repeat_lbls_y
text

facet.fixed_y
text

facet.drop
text

facet.nrow
text

facet.relative
text

x.date_breaks
text

x.date_labels
text

category.focus
text

colour
text

colour_fill
text

x.lbl_angle
text

x.lbl_align
text

x.lbl_italic
text

x.remove
text

x.position
text

x.max_items
text

x.max_txt
text

category.max_items
text

category.max_txt
text

facet.max_items
text

facet.max_txt
text

x.breaks
text

x.breaks_n
text

x.trans
text

x.expand
text

x.limits
text

x.character
text

y.remove
text

y.24h
text

y.age
text

y.percent
text

y.percent_break
text

y.breaks
text

y.limits
text

y.labels
text

y.expand
text

y.trans
text

y.position
text

category.labels
settings for the plotting direction category

category.percent
settings for the plotting direction category

category.breaks
settings for the plotting direction category

category.limits
settings for the plotting direction category

category.expand
settings for the plotting direction category

category.midpoint
settings for the plotting direction category

category.trans
settings for the plotting direction category

x.sort
sorting of the plotting direction, defaults to TRUE, except for continuous values on the x axis (such as dates and numbers). Applying one of the sorting methods will transform the values to an ordered factor, which ggplot2 uses to orient the data. Valid values are:

TRUE: sort factor on their levels, otherwise sort as "asc"

FALSE: keep order as it is in the data

NULL: do not sort/transform at all

"asc" or "alpha": sort ascending on alphabet, while maintaining numbers in the text (numeric sort)

"desc": sort descending on alphabet, while maintaining numbers in the text (numeric sort)

"order" or "inorder": sort as FALSE

"freq" or "freq-desc": sort descending according to the frequencies of y computed by summarise_function (highest value first)

"freq-asc": sort ascending according to the frequencies of y computed by summarise_function (loewest value first)

category.sort
sorting of the plotting direction, defaults to TRUE, except for continuous values on the x axis (such as dates and numbers). Applying one of the sorting methods will transform the values to an ordered factor, which ggplot2 uses to orient the data. Valid values are:

TRUE: sort factor on their levels, otherwise sort as "asc"

FALSE: keep order as it is in the data

NULL: do not sort/transform at all

"asc" or "alpha": sort ascending on alphabet, while maintaining numbers in the text (numeric sort)

"desc": sort descending on alphabet, while maintaining numbers in the text (numeric sort)

"order" or "inorder": sort as FALSE

"freq" or "freq-desc": sort descending according to the frequencies of y computed by summarise_function (highest value first)

"freq-asc": sort ascending according to the frequencies of y computed by summarise_function (loewest value first)

facet.sort
sorting of the plotting direction, defaults to TRUE, except for continuous values on the x axis (such as dates and numbers). Applying one of the sorting methods will transform the values to an ordered factor, which ggplot2 uses to orient the data. Valid values are:

TRUE: sort factor on their levels, otherwise sort as "asc"

FALSE: keep order as it is in the data

NULL: do not sort/transform at all

"asc" or "alpha": sort ascending on alphabet, while maintaining numbers in the text (numeric sort)

"desc": sort descending on alphabet, while maintaining numbers in the text (numeric sort)

"order" or "inorder": sort as FALSE

"freq" or "freq-desc": sort descending according to the frequencies of y computed by summarise_function (highest value first)

"freq-asc": sort ascending according to the frequencies of y computed by summarise_function (loewest value first)

datalabels
variables or character vector to use as datalabels - if left blank, will take the first character column in 'sf' plots, and values of y otherwise

datalabels.round
number of digits to round the datalabels

datalabels.format
format to use for datalabels - "%n" will be replaced by the count number, "%p" will be replaced by the percentage of the total count. Use datalabels.format = NULL to not transform the datalabels.

datalabels.colour
settings for the datalabels

datalabels.fill
settings for the datalabels

datalabels.size
settings for the datalabels

datalabels.angle
settings for the datalabels

decimal.mark
decimal mark, defaults to Dutch use (a comma)

big.mark
thousands separator, defaults to Dutch use (a full stop)

summarise_function
a function to use if the data has to be summarised, see Examples

stacked
text

stackedpercent
text

horizontal
a logical to turn the plot 90 defrees using coord_flip()

reverse
a logical to reverse all values on the x axis

smooth
a logical to add a smooth using geom_smooth()

smooth.method
settings for smooth

smooth.formula
settings for smooth

smooth.se
settings for smooth

smooth.level
settings for smooth

smooth.alpha
settings for smooth

smooth.size
settings for smooth

smooth.linetype
settings for smooth

size
size of the geom

linetype
linetype of the geom, only suitable for geoms that draw lines

bins
number of bins, only useful for geom = "histogram"

width
width of the geom

jitter_seed
seed (randomisation factor) to be set when using geom = "jitter"

violin_scale
scale to be set when using geom = "violin", can also be set to "area"

legend.position
settings for the legend

legend.title
settings for the legend

legend.reverse
settings for the legend

legend.barheight
settings for the legend

legend.barwidth
settings for the legend

legend.nbin
settings for the legend

legend.italic
settings for the legend

zoom
a logical to indicate if the plot should be scaled to the data, i.e., not having the x and y axes to start at 0

sep
separator character to use if multiple columns are given to either of the three directions: x, category and facet, e.g. facet = c(column1, column2)

print
a logical to indicate if the result should be printed instead of just returned

text_factor
text factor to use, which will apply to all texts shown in the plot

family
font family to use

theme
a valid ggplot2 theme to apply, or NULL to use the default theme_grey()

markdown
text

x.category
old argument for facet, now deprecated

y.category
old argument for category, now deprecated

...
arguments passed on to methods