Methods for plot2()

These are the implemented methods for different S3 classes to be used in plot2(). Since they have an extensive list of arguments, they are placed here on a separate manual page.

# S3 method for default
plot2(
  .data,
  x = NULL,
  y = NULL,
  category = NULL,
  facet = NULL,
  type = NULL,
  x.title = NULL,
  y.title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "certe",
  colour_fill = NULL,
  x.lbl_angle = 0,
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x %n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x %n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x %n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0.5,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 0.1,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0.25,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.midpoint = NULL,
  category.trans = "identity",
  x.sort = NULL,
  category.sort = TRUE,
  facet.sort = TRUE,
  datalabels = TRUE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.format = "%n",
  datalabels.colour = "grey25",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ifelse(decimal.mark == ",", ".", ","),
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = FALSE,
  horizontal = FALSE,
  reverse = horizontal,
  smooth = NULL,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = NULL,
  linetype = 1,
  binwidth = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "top",
  legend.title = NULL,
  legend.reverse = FALSE,
  legend.barheight = 6,
  legend.barwidth = 1.5,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = getOption("plot2.family"),
  theme = theme_minimal2(),
  markdown = TRUE,
  taxonomy_italic = FALSE,
  x.category = NULL,
  y.category = NULL,
  ...
)

# S3 method for numeric
plot2(
  .data,
  x = NULL,
  y = NULL,
  category = NULL,
  facet = NULL,
  type = "point",
  x.title = NULL,
  y.title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "certe",
  colour_fill = NULL,
  x.lbl_angle = 0,
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x %n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x %n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x %n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0.5,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 0.1,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0.25,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.midpoint = NULL,
  category.trans = "identity",
  x.sort = NULL,
  category.sort = TRUE,
  facet.sort = TRUE,
  datalabels = FALSE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.format = "%n",
  datalabels.colour = "grey25",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ifelse(decimal.mark == ",", ".", ","),
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = FALSE,
  horizontal = FALSE,
  reverse = horizontal,
  smooth = NULL,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = NULL,
  linetype = 1,
  binwidth = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "top",
  legend.title = NULL,
  legend.reverse = FALSE,
  legend.barheight = 6,
  legend.barwidth = 1.5,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = getOption("plot2.family"),
  theme = theme_minimal2(),
  markdown = TRUE,
  taxonomy_italic = FALSE,
  x.category = NULL,
  y.category = NULL,
  ...
)

# S3 method for freq
plot2(
  .data,
  x = .data$item,
  y = .data$count,
  category = NULL,
  facet = NULL,
  type = NULL,
  x.title = "Item",
  y.title = "Count",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "certe",
  colour_fill = NULL,
  x.lbl_angle = 0,
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x %n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x %n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x %n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0.5,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 0.1,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0.25,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.midpoint = NULL,
  category.trans = "identity",
  x.sort = "freq-desc",
  category.sort = TRUE,
  facet.sort = TRUE,
  datalabels = TRUE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.format = "%n",
  datalabels.colour = "grey25",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ifelse(decimal.mark == ",", ".", ","),
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = FALSE,
  horizontal = FALSE,
  reverse = horizontal,
  smooth = NULL,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = NULL,
  linetype = 1,
  binwidth = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "top",
  legend.title = NULL,
  legend.reverse = FALSE,
  legend.barheight = 6,
  legend.barwidth = 1.5,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = getOption("plot2.family"),
  theme = theme_minimal2(),
  markdown = TRUE,
  taxonomy_italic = FALSE,
  x.category = NULL,
  y.category = NULL,
  ...
)

# S3 method for sf
plot2(
  .data,
  x = NULL,
  y = NULL,
  category = NULL,
  facet = NULL,
  type = NULL,
  x.title = NULL,
  y.title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "grey25",
  colour_fill = c("white", "certeblauw"),
  x.lbl_angle = 0,
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x %n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x %n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x %n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 0.1,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.midpoint = NULL,
  category.trans = "identity",
  x.sort = NULL,
  category.sort = TRUE,
  facet.sort = TRUE,
  datalabels = TRUE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.format = NULL,
  datalabels.colour = "black",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ifelse(decimal.mark == ",", ".", ","),
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = FALSE,
  horizontal = FALSE,
  reverse = horizontal,
  smooth = NULL,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = 0.1,
  linetype = 1,
  binwidth = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "right",
  legend.title = TRUE,
  legend.reverse = FALSE,
  legend.barheight = 6,
  legend.barwidth = 1.5,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = getOption("plot2.family"),
  theme = theme_minimal2(panel.grid.major = element_blank(), panel.grid.minor =
    element_blank(), panel.border = element_blank(), axis.title = element_blank(),
    axis.text = element_blank(), axis.line = element_blank(), axis.ticks =
    element_blank()),
  markdown = TRUE,
  taxonomy_italic = FALSE,
  x.category = NULL,
  y.category = NULL,
  ...
)

# S3 method for data.frame
plot2(
  .data,
  x = NULL,
  y = NULL,
  category = NULL,
  facet = NULL,
  type = NULL,
  x.title = NULL,
  y.title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "certe",
  colour_fill = NULL,
  x.lbl_angle = 0,
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x %n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x %n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x %n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0.5,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 0.1,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0.25,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.midpoint = NULL,
  category.trans = "identity",
  x.sort = NULL,
  category.sort = TRUE,
  facet.sort = TRUE,
  datalabels = TRUE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.format = "%n",
  datalabels.colour = "grey25",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ifelse(decimal.mark == ",", ".", ","),
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = FALSE,
  horizontal = FALSE,
  reverse = horizontal,
  smooth = NULL,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = NULL,
  linetype = 1,
  binwidth = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "top",
  legend.title = NULL,
  legend.reverse = FALSE,
  legend.barheight = 6,
  legend.barwidth = 1.5,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = getOption("plot2.family"),
  theme = theme_minimal2(),
  markdown = TRUE,
  taxonomy_italic = FALSE,
  x.category = NULL,
  y.category = NULL,
  ...
)

# S3 method for bug_drug_combinations
plot2(
  .data,
  x = ab,
  y = value,
  category = name,
  facet = mo,
  type = "column",
  x.title = "",
  y.title = "",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "certe_rsi2",
  colour_fill = NULL,
  x.lbl_angle = ifelse(horizontal, 0, 90),
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x %n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x %n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x %n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0.5,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 0.1,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0.25,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.midpoint = NULL,
  category.trans = "identity",
  x.sort = NULL,
  category.sort = FALSE,
  facet.sort = TRUE,
  datalabels = TRUE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.format = "%n",
  datalabels.colour = "grey25",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ifelse(decimal.mark == ",", ".", ","),
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = TRUE,
  horizontal = TRUE,
  reverse = TRUE,
  smooth = NULL,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = NULL,
  linetype = 1,
  binwidth = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "top",
  legend.title = NULL,
  legend.reverse = FALSE,
  legend.barheight = 6,
  legend.barwidth = 1.5,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = getOption("plot2.family"),
  theme = theme_minimal2(),
  markdown = TRUE,
  taxonomy_italic = TRUE,
  x.category = NULL,
  y.category = NULL,
  minimum = 30,
  remove_intrinsic_resistant = TRUE,
  language = "nl",
  ...
)

Arguments

.data
data to plot

x
plotting 'direction': the x axis

y
values to use for plotting along the y axis

category
plotting 'direction': the category (called 'fill' and 'colour' in ggplot2)

facet
plotting 'direction': the facet

type
type of visualisation to use. This can be:

A ggplot2 geom name, all geoms are supported. Full function names can be used (e.g., "geom_histogram"), but they can also be abbreviated (e.g., "h", "hist"). These geoms can be abbreviated by their first character: area ("a"), boxplot ("b"), column ("c"), histogram ("h"), jitter ("j"), line ("l"), point ("p"), ribbon ("r"), violin ("v"). Please note: in ggplot2, 'bars' and 'columns' are equal, while it is common to many people that 'bars' are oriented horizontally and 'columns' are oriented vertically. For this reason, type = "bar" will set type = "col" and horizontal = TRUE.

A shortcut. There is currently one supported shortcut: "barpercent", which will set type = "col" and horizontal = TRUE and x.max_items = 10 and x.sort = "freq-desc" and datalabels.format = "%n (%p)".

Left blank. In this case, the type will be determined automatically: "boxplot" if there is no X axis or if the length of unique values per X axis item is at least 3, "point" if both the Y and X axes are numeric, and "col" otherwise.

x.title
text to show on the x axis

y.title
text to show on the y axis

title
title to show

subtitle
subtitle to show

caption
caption to show

tag
tag to show

title.linelength
maximum number of characters per line in the title, before a linebreak occurs

title.colour
text colour of the title

subtitle.linelength
maximum number of characters per line in the subtitle, before a linebreak occurs

subtitle.colour
text colour of the subtitle

na.replace
character to put in place of NA values if na.rm = FALSE

na.rm
remove NA values from showing in the plot

facet.position
settings for the plotting direction facet

facet.fill
settings for the plotting direction facet

facet.bold
settings for the plotting direction facet

facet.italic
settings for the plotting direction facet

facet.size
settings for the plotting direction facet

facet.margin
settings for the plotting direction facet

facet.repeat_lbls_x
settings for the plotting direction facet

facet.repeat_lbls_y
settings for the plotting direction facet

facet.fixed_y
settings for the plotting direction facet

facet.drop
settings for the plotting direction facet

facet.nrow
settings for the plotting direction facet

facet.relative
settings for the plotting direction facet

x.date_breaks
breaks to use when the x axis contains dates, will be determined automatically if left blank

x.date_labels
labels to use when the x axis contains dates, will be determined automatically if left blank

category.focus
a value of category that should be highlighted, meaning that all other values in category will be greyed out. This can also be a numeric value between 1 and the length of unique values of category, e.g. category.focus = 2 to focus on the second legend item.

colour
colour(s) to set, will be evaluated with colourpicker() and defaults to Certe colours. This can be a named vector to match values of category, see Examples. Using a named vector can also be used to manually sort the values of category.

colour_fill
colour(s) to be used for filling, will be determined automatically if left blank and will be evaluated with colourpicker()

x.lbl_angle
angle to use for the x axis in a counter-clockwise direction (i.e., a value of 90 will orient the axis labels from bottom to top, a value of 270 will orient the axis labels from top to bottom)

x.lbl_align
alignment for the x axis between 0 (left aligned) and 1 (right aligned)

x.lbl_italic
a logical to indicate whether the x labels should in in italics

x.remove
a logical to indicate whether the x labels and title should be removed

x.position
position of the x axis, defaults to "bottom"

x.max_items
number of maximum items to use, defaults to infinite. All other values will be grouped and summarised using the summarise_function function. Please note: the sorting will be applied first, allowing to e.g. plot the top n most frequent values of the x axis by combining x.sort = "freq-desc" with x.max_items = n.

x.max_txt
the text to use of values not included number of *.max_items. The placeholder %n will be replaced with the outcome of the summarise_function function, the placeholder %p will be replaced with the percentage.

category.max_items
number of maximum items to use, defaults to infinite. All other values will be grouped and summarised using the summarise_function function. Please note: the sorting will be applied first, allowing to e.g. plot the top n most frequent values of the x axis by combining x.sort = "freq-desc" with x.max_items = n.

category.max_txt
the text to use of values not included number of *.max_items. The placeholder %n will be replaced with the outcome of the summarise_function function, the placeholder %p will be replaced with the percentage.

facet.max_items
number of maximum items to use, defaults to infinite. All other values will be grouped and summarised using the summarise_function function. Please note: the sorting will be applied first, allowing to e.g. plot the top n most frequent values of the x axis by combining x.sort = "freq-desc" with x.max_items = n.

facet.max_txt
the text to use of values not included number of *.max_items. The placeholder %n will be replaced with the outcome of the summarise_function function, the placeholder %p will be replaced with the percentage.

x.breaks
a breaks function or numeric vector to use for the x axis

x.breaks_n
number of breaks to use for the x axis

x.trans
a transformation function to use for the x axis, e.g. "log2"

x.expand
expansion to use for the x axis, can be length 1 or 2

x.limits
limits to use for the x axis, can be length 1 or 2

x.character
a logical to indicate whether the values of the x axis should be forced to character. The default is FALSE, except for years (x values between 2000 and 2050)

y.remove
a logical to indicate whether the y labels and title should be removed

y.24h
a logical to indicate whether the y labels and breaks should be formatted as 24-hour sequences

y.age
a logical to indicate whether the y labels and breaks should be formatted as ages in years

y.percent
a logical to indicate whether the y labels should be formatted as percentages

y.percent_break
number of percentages on which the y axis should have breaks

y.breaks
a breaks function or numeric vector to use for the y axis

y.limits
limits to use for the y axis, can be length 1 or 2

y.labels
a labels function or character vector to use for the y axis

y.expand
expansion to use for the y axis, can be length 1 or 2

y.trans
a transformation function to use for the y axis, e.g. "log2"

y.position
position of the x axis, defaults to "left"

category.labels
settings for the plotting direction category

category.percent
settings for the plotting direction category

category.breaks
settings for the plotting direction category

category.limits
settings for the plotting direction category

category.expand
settings for the plotting direction category

category.midpoint
settings for the plotting direction category

category.trans
settings for the plotting direction category

x.sort
sorting of the plotting direction, defaults to TRUE, except for continuous values on the x axis (such as dates and numbers). Applying one of the sorting methods will transform the values to an ordered factor, which ggplot2 uses to orient the data. Valid values are:

TRUE: sort as "asc"

FALSE: sort according to the order in the data

NULL: do not sort/transform at all

"asc" or "alpha": sort factors on their levels, otherwise sort ascending on alphabet, while maintaining numbers in the text (numeric sort)

"desc": sort factors on their reversed levels, otherwise sort descending on alphabet, while maintaining numbers in the text (numeric sort)

"order" or "inorder": sort as FALSE

"freq" or "freq-desc": sort descending according to the frequencies of y computed by summarise_function (highest value first)

"freq-asc": sort ascending according to the frequencies of y computed by summarise_function (loewest value first)

category.sort
sorting of the plotting direction, defaults to TRUE, except for continuous values on the x axis (such as dates and numbers). Applying one of the sorting methods will transform the values to an ordered factor, which ggplot2 uses to orient the data. Valid values are:

TRUE: sort as "asc"

FALSE: sort according to the order in the data

NULL: do not sort/transform at all

"asc" or "alpha": sort factors on their levels, otherwise sort ascending on alphabet, while maintaining numbers in the text (numeric sort)

"desc": sort factors on their reversed levels, otherwise sort descending on alphabet, while maintaining numbers in the text (numeric sort)

"order" or "inorder": sort as FALSE

"freq" or "freq-desc": sort descending according to the frequencies of y computed by summarise_function (highest value first)

"freq-asc": sort ascending according to the frequencies of y computed by summarise_function (loewest value first)

facet.sort
sorting of the plotting direction, defaults to TRUE, except for continuous values on the x axis (such as dates and numbers). Applying one of the sorting methods will transform the values to an ordered factor, which ggplot2 uses to orient the data. Valid values are:

TRUE: sort as "asc"

FALSE: sort according to the order in the data

NULL: do not sort/transform at all

"asc" or "alpha": sort factors on their levels, otherwise sort ascending on alphabet, while maintaining numbers in the text (numeric sort)

"desc": sort factors on their reversed levels, otherwise sort descending on alphabet, while maintaining numbers in the text (numeric sort)

"order" or "inorder": sort as FALSE

"freq" or "freq-desc": sort descending according to the frequencies of y computed by summarise_function (highest value first)

"freq-asc": sort ascending according to the frequencies of y computed by summarise_function (loewest value first)

datalabels
variables or character vector to use as datalabels - if left blank, will take the first character column in 'sf' plots, and values of y otherwise

datalabels.round
number of digits to round the datalabels

datalabels.format
format to use for datalabels - "%n" will be replaced by the count number, "%p" will be replaced by the percentage of the total count. Use datalabels.format = NULL to not transform the datalabels.

datalabels.colour
settings for the datalabels

datalabels.fill
settings for the datalabels

datalabels.size
settings for the datalabels

datalabels.angle
settings for the datalabels

decimal.mark
decimal mark, defaults to Dutch use (a comma)

big.mark
thousands separator, defaults to Dutch use (a full stop)

summarise_function
a function to use if the data has to be summarised, see Examples

stacked
a logical to indicate that values must be stacked

stackedpercent
a logical to indicate that values must be 100% stacked

horizontal
a logical to turn the plot 90 degrees using coord_flip(). This option also updates some theme options, so that e.g., x.lbl_italic will still apply to the original x axis.

reverse
a logical to reverse all values of category

smooth
a logical to add a smooth. In histograms, this will add the density count as an overlaying line (default: TRUE). In all other cases, a smooth will be added using geom_smooth() (default: FALSE).

smooth.method
settings for smooth

smooth.formula
settings for smooth

smooth.se
settings for smooth

smooth.level
settings for smooth

smooth.alpha
settings for smooth

smooth.size
settings for smooth

smooth.linetype
settings for smooth

size
size of the geom

linetype
linetype of the geom, only suitable for geoms that draw lines

binwidth
width of bins (only useful for geom = "histogram"), can be specified as a numeric value or as a function that calculates width from x, see geom_histogram(). It defaults to approx. diff(range(x)) divided by 12 to 22 based on the data.

width
width of the geom

jitter_seed
seed (randomisation factor) to be set when using type = "jitter"

violin_scale
scale to be set when using type = "violin", can also be set to "area"

legend.position
settings for the legend

legend.title
settings for the legend

legend.reverse
settings for the legend

legend.barheight
settings for the legend

legend.barwidth
settings for the legend

legend.nbin
settings for the legend

legend.italic
settings for the legend

zoom
a logical to indicate if the plot should be scaled to the data, i.e., not having the x and y axes to start at 0

sep
separator character to use if multiple columns are given to either of the three directions: x, category and facet, e.g. facet = c(column1, column2)

print
a logical to indicate if the result should be printed instead of just returned

text_factor
text factor to use, which will apply to all texts shown in the plot

family
font family to use, can be set with options(plot2.family = "...")

theme
a valid ggplot2 theme to apply, or NULL to use the default theme_grey()

markdown
a logical to turn all labels and titles into markdown-supported labels, by extending their S3 classes with "element_markdown", a feature of the ggtext package

taxonomy_italic
a logical to transform all labels and titles into italics that are in the microorganisms data set of the AMR package

x.category
old argument for facet, now deprecated

y.category
old argument for category, now deprecated

...
arguments passed on to methods

minimum
minimum number of results, defaults to 30

remove_intrinsic_resistant
a logical to indicate that rows with 100% resistance must be removed from the data set before plotting

language
language to be used for antibiotic names

Details

For geographic information system (GIS) analysis, use the sf package with a data set containing geometries. The result can be used as input for plot2().

For antimicrobial resistance (AMR) data analysis, use the bug_drug_combinations() function from the AMR package on a data set with antibiograms. The result can be used as input for plot2().